add enterroom and exitroom lobby information broadcast
