+++ src/main/java/ch/uzh/ifi/hase/soprafs24/model/Response.java
@@ -1,0 +1,31 @@
+package ch.uzh.ifi.hase.soprafs24.model;
+
+public class Response {
+    private boolean success;
+    private String receipt;
+    private String message;
+
+    public String getReceipt() {
+        return receipt;
+    }
+
+    public void setReceipt(String receipt) {
+        this.receipt = receipt;
+    }
+    
+    public boolean isSuccess() {
+        return success;
+    }
+
+    public void setSuccess(boolean success) {
+        this.success = success;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
