--- src/main/java/ch/uzh/ifi/hase/soprafs24/model/TimestampedResponse.java
@@ -1,22 +1,0 @@
-package ch.uzh.ifi.hase.soprafs24.model;
-
-public class TimestampedResponse {
-    private long timestamp;
-    private String message;
-
-    public long getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
